New action on RP, only render not commit
This new Github Action will be triggered when there is a PR from anywhere
where the docs/ directory is changed.
This action will only test if the docs render correctly, but it will not try to
publish the rendered files on gh-pages. This will only happen on push to master
(which means when the PR is merged with master)